Server Authentication From Internet Security and Privacy by Boris
Last meeting we focused on authentication of a person to a server or to another peer. In the next meeting the focus will be about authentication of a server or a business to a user.

This event will be hosted either by Rival Technologies or Lighthouse Labs. Check this page for an update later on. I have set Rival Technologies for now.

Agenda: These topics will de discussed:

- HTTPS protocol and Certificate Authorities
- Mass Surveillance
- DomainKeys (email)
- DNSSEC
- TACK protocol (by Moxie Marlinspike)
- Secure Remote Password protocol (SRP)
- OPAQUE protocol
- Elliptic Curve Cryptography if time permits.